backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
psycopg2-binary==2.9.1
dj-database-url==0.5.0
python-dotenv==0.20.0
django-environ==0.11.2
